move _legacy_evaluate to evaluation/base.py and replace prints with logging
